+++ src/main/java/org/magicdgs/readtools/tools/barcodes/dictionary/BarcodeDictionary.java
@@ -168,7 +168,7 @@ public class BarcodeDictionary {
      * @return the sample names
      */
     public List<SAMReadGroupRecord> getSampleReadGroups() {
-        return sampleRecord;
+        return Collections.unmodifiableList(sampleRecord);
     }
 
     /**